manual: also install images/*.png with the HTML output
Extend the CMake rules for the manual. Do install the images/ subdir but
omit the *.xcf files. Prepare the CSS file although this seems to be not
strictly necessary (gets inlined, but we are prepared if it's external).
Only install these additional files when the asciidoctor(1) tool is
available, and thus the .html file can get generated (to not install
some questionable combination of a .pdf and unrelated .png files, or to
not install the resources when none of the manual files got created).

This approach was tested with the following command sequence:
  $ make manual
  $ make install
  $ xdg-open ~/share/doc/pulseview/*.html