Show relative time of flags on hover
With this change, when the mouse hovers over any flag or cursor, the
flags change labels to show their time relative to the item being
hovered over. When hovering over the interval label of a cursor pair,
the relative time to the nearest end of the interval is shown. While
pressing shift, the regular labels are shown again.