TraceView: Prevent decode traces from changing order
Before, decode traces sometimes switched places when opening a
class selector because max_visible_rows was reset to 0 which
makes v_extents() return a value so small that the next decode
trace would move up.

Also, triggering the lazy event handler was changed so that it
fires even when it's triggered several times. Before, it would
always reset and only fire once, after the triggering stopped.
That is undesirable when a class selector opens.