cmake: move project() before other statements
The included GNUInstallDirs logic needs to know about the programming
languages which are used in the project. Without this spec a verbose
developer message gets emitted:

  CMake Warning (dev) at $HOME/share/cmake-3.18/Modules/GNUInstallDirs.cmake:225 (message):
    Unable to determine default CMAKE_INSTALL_LIBDIR directory because no
    target architecture is known.  Please enable at least one language before
    including GNUInstallDirs.
  Call Stack (most recent call first):
    CMakeLists.txt:24 (include)
    This warning is for project developers.  Use -Wno-dev to suppress it.

Move the project() statement in CMakeList.txt before the include(), and
specify the C and C++ programming languages (C is needed at configuration
time for feature detection, before the application's C++ sources get built).