Created new logic math signal
Added a logic math signal so that logic signals can be combined using an
expression. This can be particularly useful when wanting to visually
identify interesting locations to investigate within a series of logic
traces.

This was accomplished by refactoring the math signal into mostly generic
code with descendants that flesh out the analog and logic specific code.