Signal: Add missing item separator in channel names list
The SCL item in the ChannelNames[] table lacked the separator, which
made it "run into" the next item (preprocessor string concatenation).
Add the missing separator.

This amends commits 9e40e83daf6a (which introduced the lack of the
separator while it did not take effect yet) and 7d5a9c3e79cb (which
made it take effect when items got added and rearranged).

Reported-By: Dominik Sliwa <dominik@sliwa.io>