UI: Move shortcuts to Qt StandardKeys constants
This patch uses the Qt provided constants for the quit, close,
open and save keyboard shortcuts. This removes few instances of
Qt version specific code.

Only functional difference of this patch is that on windows the
quit shortcut of ctrl-q will no longer work, instead relying on
the platform standard alt-f4.